Refactoring: class GeoDataPointmap::DeletePointsCommand moved to independent files.
