+++ libs/geodata/pointmap/pointmap.pro
@@ -156,6 +156,7 @@ HEADERS += gd_pointmap_global.h \
            geodatapointmapwebimportersetting.h \
            geodatapointmapwebimportersettingmanager.h \
            geodatapointmapwebimporterzoomleveldialog.h \
+           private/geodatapointmap_deletepointscommand.h \
            private/geodatapointmapcellmappert_detail.h \
            private/geodatapointmapnodemappert_detail.h \
            private/geodatapointmapt_detail.h \
@@ -191,7 +192,8 @@ SOURCES += geodatapointmap.cpp \
            geodatapointmapwebimporterregionselectdialog.cpp \
            geodatapointmapwebimportersetting.cpp \
            geodatapointmapwebimportersettingmanager.cpp \
-           geodatapointmapwebimporterzoomleveldialog.cpp
+           geodatapointmapwebimporterzoomleveldialog.cpp \
+           private/geodatapointmap_deletepointscommand.cpp
 RESOURCES += pointmap.qrc
 TRANSLATIONS += languages/iricGdPointmap_es_ES.ts \
                 languages/iricGdPointmap_fr_FR.ts \
